🚚 Move CustomHeightTabBar class to CommonUI Group
