+++ HRHN/Sources/UI/TabBar/TabBarController.swift
@@ -5,18 +5,7 @@
 //  Created by Chanhee Jeong on 2022/12/20.
 //
 
-// https://stackoverflow.com/questions/48192749/hidesbottombarwhenpushed-makes-uitabbar-jump
-// https://gist.github.com/calt/7ea29a65b440c2aa8a1a
-
 import UIKit
-
-class CustomHeightTabBar : UITabBar {
-    override open func sizeThatFits(_ size: CGSize) -> CGSize {
-        var sizeThatFits = super.sizeThatFits(size)
-        sizeThatFits.height = 99
-        return sizeThatFits
-    }
-}
 
 class TabBarController: UITabBarController {
